Move functions about audio to audio.cpp
