+++ CMakeLists.txt
@@ -43,6 +43,7 @@ set(CORE_SOURCES
     access_character_info.cpp
     access_item_db.cpp
     animation.cpp
+    audio.cpp
     autopick.cpp
     buff.cpp
     calc.cpp
