+++ std.cpp
@@ -1792,144 +1792,6 @@ void memcpy_(
 
 
 
-namespace mixer_detail
-{
-
-std::unordered_map<int, Mix_Chunk*> chunks;
-}
-
-
-
-int DSINIT()
-{
-    Mix_AllocateChannels(16);
-    snail::application::instance().register_finalizer([&]() {
-        for (const auto& pair : mixer_detail::chunks)
-        {
-            if (pair.second)
-                ::Mix_FreeChunk(pair.second);
-        }
-    });
-    return 1;
-}
-
-
-
-void DSEND()
-{
-}
-
-
-
-void DSRELEASE(int)
-{
-}
-
-
-
-void DSLOADFNAME(const fs::path& filepath, int channel)
-{
-    if (mixer_detail::chunks.find(channel) != std::end(mixer_detail::chunks))
-    {
-        if (mixer_detail::chunks[channel])
-            Mix_FreeChunk(mixer_detail::chunks[channel]);
-    }
-    auto chunk = snail::detail::enforce_mixer(
-        Mix_LoadWAV(filesystem::to_utf8_path(filepath).c_str()));
-    mixer_detail::chunks[channel] = chunk;
-}
-
-
-
-void DSPLAY(int channel, int loop)
-{
-    Mix_PlayChannel(-1, mixer_detail::chunks[channel], loop ? -1 : 0);
-}
-
-
-
-void DSSTOP(int channel)
-{
-    Mix_HaltChannel(channel);
-}
-
-
-
-void DSSETVOLUME(int, int)
-{
-}
-
-
-
-int DSGETMASTERVOLUME()
-{
-    return 100;
-}
-
-
-
-int CHECKPLAY(int channel)
-{
-    return Mix_Playing(channel) ? 1 : 0;
-}
-
-
-
-namespace mixer_detail
-{
-
-Mix_Music* music = nullptr;
-}
-
-
-
-int DMINIT()
-{
-    snail::application::instance().register_finalizer([&]() {
-        if (mixer_detail::music)
-            ::Mix_FreeMusic(mixer_detail::music);
-    });
-    return 1;
-}
-
-
-
-void DMEND()
-{
-}
-
-
-void DMLOADFNAME(const fs::path& filepath, int)
-{
-    if (mixer_detail::music)
-        ::Mix_FreeMusic(mixer_detail::music);
-
-    mixer_detail::music = snail::detail::enforce_mixer(
-        Mix_LoadMUS(filesystem::to_utf8_path(filepath).c_str()));
-}
-
-
-
-void DMPLAY(int loop, int)
-{
-    snail::detail::enforce_mixer(
-        Mix_PlayMusic(mixer_detail::music, loop ? -1 : 1));
-}
-
-
-
-void DMSTOP()
-{
-    ::Mix_HaltMusic();
-    if (mixer_detail::music)
-    {
-        ::Mix_FreeMusic(mixer_detail::music);
-        mixer_detail::music = nullptr;
-    }
-}
-
-
-
 void DIINIT()
 {
 }
